requirements.txt
requests==2.13.0
stitchstream-python==0.5.0
backoff==1.3.2
